+++ bin/commands/cpp/cmakelists.d.ts
@@ -1,1 +1,0 @@
-export declare function getTemplateCMakeLists(project_name: string, header_files: string[], source_files: string[], config: any): string;