+++ bin/commands/cpp/cmakelists.test.d.ts
@@ -1,1 +1,0 @@
-export declare function getTemplateCMakeListsTest(test_project_name: string, headers: string[], source: string[], config: any): string;