+++ bin/commands/cpp/cmakelists.test.js
@@ -41,4 +41,3 @@ target_link_libraries( test.${test_project_name} \${LIB_FILES} \${CMAKE_THREAD_L
     return cmakelists_test;
 }
 exports.getTemplateCMakeListsTest = getTemplateCMakeListsTest;
-//# sourceMappingURL=../../../src/commands/cpp/cmakelists.test.js.map