+++ bin/commands/cpp/header.d.ts
@@ -1,1 +1,0 @@
-export declare function getTemplateHeader(filename: string, config: any): string;