+++ bin/commands/cpp/main.d.ts
@@ -1,1 +1,0 @@
-export declare function getTemplateMain(header_files: string[], options: any): string;