+++ bin/commands/cpp/source.d.ts
@@ -1,1 +1,0 @@
-export declare function getTemplateSource(filename: string, options: any): string;