+++ bin/commands/cpp/source.js
@@ -4,8 +4,7 @@ exports.getTemplateSource = void 0;
 const title_1 = require("./title");
 function getTemplateSource(filename, options) {
     const header_file = filename.replace(/\.cpp$/i, '.hpp');
-    const source = `${title_1.getTemplateTitle(filename, options)}\n\n#include "${header_file}"\n`;
+    const source = `${(0, title_1.getTemplateTitle)(filename, options)}\n\n#include "${header_file}"\n`;
     return source;
 }
 exports.getTemplateSource = getTemplateSource;
-//# sourceMappingURL=../../../src/commands/cpp/source.js.map