+++ bin/commands/cpp/title.d.ts
@@ -1,1 +1,0 @@
-export declare function getTemplateTitle(filename: string, config: any): string;