+++ bin/commands/create-cpp-project.d.ts
@@ -1,1 +1,0 @@
-export declare function createCppProject(cmd: any, options: any): void;