+++ bin/commands/create-js-project.d.ts
@@ -1,1 +1,0 @@
-export declare function createJSProject(cmd: any, options: any): void;