+++ bin/lib/utils.d.ts
@@ -1,3 +1,0 @@
-export declare const stub_test: () => string;
-export declare function downloadFileHttps(uri: any, filename: string, cb: any): void;
-export declare function downloadFileHttp(uri: string, filename: string, cb: any): void;