+++ ry.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+npm run build
+npm pack
+npm i -g dm-tools-0.4.0.tgz