+++ src/main.ts
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 'use strict';
-import * as cmd from 'commander';
+import { Command } from 'commander';
 import { VERSION } from './data-types/data-types';
 import * as fs from 'fs';
 
 import { createJSProject } from './commands/create-js-project';
-import { createCppProject } from './commands/create-cpp-project';
+// import { createCppProject } from './commands/create-cpp-project';
 
 const options: any = {};
 const config_file = process.env.HOME + '/dmtools.json';
@@ -25,14 +25,17 @@ function source_files( files: string ) {
   }
 }
 */
+const cmd = new Command();
+
 let config: any = { project: {} };
 if ( fs.existsSync( config_file ) ) {
   config = JSON.parse( fs.readFileSync( config_file, 'utf8' ) );
 }
 Object.assign( options, { config } );
 
+
 cmd
-  .version( `DevMentor Tools Project Generator v${ VERSION }`, '-v, --version' )
+  .version( `DevMentor Tools Project Generator v${ VERSION }`, '-v, --version', 'Show current version' )
   .usage( '<command> <project> [options...]\n\n     commands: new' )
   .arguments( '<command> <project>' )
   .option( '-t, --type <type>', 'project types: gql, js, koa, node, ts, web' )
@@ -56,14 +59,16 @@ Website: https://www.npmjs.com/package/dm-tools
   } )
   .parse( process.argv );
 
-switch ( options.command ) {
-  case 'new':
-    if ( cmd.cpp ) {
-      createCppProject( cmd, options );
-    } else {
-      createJSProject( cmd, options );
-    }
-    break;
-  default:
-    console.log( 'Unknown Command, doing nothing!' );
-}
+createJSProject( cmd.opts(), options );
+
+// switch ( options.command ) {
+// case 'new':
+//   if ( options.cpp ) {
+//     createCppProject( options );
+//   } else {
+//     createJSProject( options );
+//   }
+// break;
+//   default:
+// console.log( 'Unknown Command, doing nothing!' );
+// }